feat: Adicionar fluxo de trabalho do GitHub Actions para o Maven CI, modernização geral do codigo

+++ src/main/java/br/nubank/adapters/json/dto/TaxDTO.java
@@ -1,3 +1,5 @@
 package br.nubank.adapters.json.dto;
 
-public record TaxDTO(double tax) {}
+import java.math.BigDecimal;
+
+public record TaxDTO(BigDecimal tax) {}
